Move stack classes to their own dedicated package
