+++ src/test/java/com/kgarsjo/datastructs/tests/EasyStackTest.java
@@ -1,6 +1,6 @@
 package com.kgarsjo.datastructs.tests;
 
-import com.kgarsjo.datastructs.EasyStack;
+import com.kgarsjo.datastructs.stack.EasyStack;
 import org.junit.Before;
 
 public class EasyStackTest extends AbstractStackTest {
