+++ src/test/java/com/kgarsjo/datastructs/tests/LinkedStackTest.java
@@ -1,6 +1,6 @@
 package com.kgarsjo.datastructs.tests;
 
-import com.kgarsjo.datastructs.LinkedStack;
+import com.kgarsjo.datastructs.stack.LinkedStack;
 import org.junit.Before;
 
 public class LinkedStackTest extends AbstractStackTest {
